Update pythran to match the paper
this is quick'n dirty and should *never* make its way to master

+++ doc/papers/wpmvp14/experiments/euler06.c
@@ -1,6 +1,6 @@
-unsigned int solve(unsigned int n) {
-  unsigned int a = 0, s = 0;
-  for(unsigned int i = 0; i< n; ++i) {
+unsigned int solve(long n) {
+  long a = 0, s = 0;
+  for(long i = 0; i< n; ++i) {
     a += i;
     s += i*i;
   }
